pyproject.toml
[project]
name = "energy-bench"
version = "0.1.0"
description = "A novel LLM evaluation benchmark focusing on energy efficient and correct code generation"
authors = [
    {name = "diones23",email = "diones23@student.aau.dk"}
]
package-mode = false
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "ollama (>=0.4.7,<0.5.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "numpy (>=2.2.4,<3.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "plotly (>=6.0.1,<7.0.0)",
    "openai (>=1.74.0,<2.0.0)",
]

[project.scripts]
energy-bench = "__main__:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
